:green_heart: Fix imports in tests after moving model/factories
