+++ src/bptl/camunda/tests/test_execute_task.py
@@ -10,8 +10,8 @@ from django_camunda.utils import serialize_variable
 from requests.exceptions import ConnectionError
 
 from bptl.tasks.models import TaskMapping
+from bptl.tasks.tests.factories import DefaultServiceFactory
 from bptl.utils.constants import Statuses
-from bptl.work_units.zgw.tests.factories import DefaultServiceFactory
 from bptl.work_units.zgw.tests.utils import mock_service_oas_get
 
 from .factories import ExternalTaskFactory
