+++ src/bptl/work_units/zgw/tests/test_eigenschappen.py
@@ -4,8 +4,7 @@ import requests_mock
 from django_camunda.utils import serialize_variable
 
 from bptl.camunda.models import ExternalTask
-from bptl.tasks.tests.factories import TaskMappingFactory
-from bptl.work_units.zgw.tests.factories import DefaultServiceFactory
+from bptl.tasks.tests.factories import DefaultServiceFactory, TaskMappingFactory
 
 from ..tasks import CreateEigenschap
 from .utils import mock_service_oas_get
