+++ src/bptl/work_units/zgw/tests/test_log_client.py
@@ -6,8 +6,7 @@ import requests_mock
 from zgw_consumers.constants import APITypes
 
 from bptl.camunda.tests.factories import ExternalTaskFactory
-from bptl.tasks.tests.factories import TaskMappingFactory
-from bptl.work_units.zgw.tests.factories import DefaultServiceFactory
+from bptl.tasks.tests.factories import DefaultServiceFactory, TaskMappingFactory
 
 from ..tasks.base import ZGWWorkUnit
 from .utils import mock_service_oas_get
