+++ src/bptl/work_units/zgw/tests/test_task_create_status.py
@@ -5,8 +5,7 @@ from django_camunda.utils import serialize_variable
 from freezegun import freeze_time
 
 from bptl.camunda.models import ExternalTask
-from bptl.tasks.tests.factories import TaskMappingFactory
-from bptl.work_units.zgw.tests.factories import DefaultServiceFactory
+from bptl.tasks.tests.factories import DefaultServiceFactory, TaskMappingFactory
 
 from ..tasks import CreateStatusTask
 from .utils import mock_service_oas_get
